+++ tests/test_date_parser.py
@@ -12,93 +12,182 @@ from nose_parameterized import parameterized, param
 import dateparser.timezone_parser
 from dateparser.date import DateDataParser, date_parser
 from dateparser.date_parser import DateParser
-from dateparser.languages import LanguageDataLoader
+from dateparser.languages import default_language_loader
 from dateparser.languages.detection import AutoDetectLanguage, ExactLanguages
 from dateparser.conf import settings
 
 from tests import BaseTestCase
 
 
-class AutoDetectLanguageTest(unittest.TestCase):
-
+class AutoDetectLanguageTest(BaseTestCase):
     def setUp(self):
-        self.parser = AutoDetectLanguage()
+        super(AutoDetectLanguageTest, self).setUp()
 
-    def test_detect_language(self):
-        self.assertItemsEqual(['es', 'pt'],
-                              map(attrgetter('shortname'), self.parser.iterate_applicable_languages('11 abril 2010')))
-        self.assertItemsEqual(['es'],
-                              map(attrgetter('shortname'), self.parser.iterate_applicable_languages('11 junio 2010')))
+        # Just a known subset so we can rely on test outcomes. Feel free to add, but not exclude or change order.
+        self.known_languages = ['en', 'fr', 'es', 'pt', 'ru', 'tr', 'cz']
 
-    @unittest.skip('This test should only be testing detecting languages, not parsing them. Although tests '
-                   'for parsing this dates should be created separately to not reduce the coverage')
-    def test_should_reduce_possible_languages_and_reject_different(self):
-        dates_in_spanish = [
-            (u'13 Ago, 2014', datetime(2014, 8, 13)),
-            (u'13 Septiembre, 2014', datetime(2014, 9, 13)),
-        ]
+        self.parser = NotImplemented
+        self.detected_languages = NotImplemented
 
-        for date_string, correct_date in dates_in_spanish:
-            parsed_date = self.parser.parse(date_string, None)
-            self.assertEqual(correct_date.date(), parsed_date.date())
+    @parameterized.expand([
+        param(date_strings=["11 abril 2010"], expected_languages=['es', 'pt']),
+        param(date_strings=["11 junio 2010"], expected_languages=['es']),
+        param(date_strings=["13 Ago, 2014", "13 Septiembre, 2014"], expected_languages=['es']),
+    ])
+    def test_detect_languages(self, date_strings, expected_languages):
+        self.given_parser(languages=self.known_languages)
+        self.when_all_languages_are_detected(date_strings)
+        self.then_detected_languages_are(expected_languages)
 
-        with self.assertRaisesRegexp(ValueError, 'Invalid date'):
-            portuguese_date = u'13 Setembro, 2014'
-            self.parser.parse(portuguese_date, None)
+    @parameterized.expand([
+        param(date_strings=["11 abril 2010"], expected_language='es'),
+        param(date_strings=["11 junio 2010"], expected_language='es'),
+        param(date_strings=["13 Ago, 2014", "13 Septiembre, 2014"], expected_language='es'),
+    ])
+    def test_exclude_ineligible_languages_with_modify(self, date_strings, expected_language):
+        self.given_parser(languages=self.known_languages)
+        self.when_one_language_is_detected(date_strings, modify=True)
+        self.then_detected_languages_are([expected_language])
+        self.then_parser_languages_are(self.known_languages[self.known_languages.index(expected_language):])
 
-    @unittest.skip('This test should only be testing detecting languages, not parsing them. Although tests '
-                   'for parsing this dates should be created separately to not reduce the coverage')
-    def test_should_accept_dates_in_different_languages(self):
-        date_fixtures = [
-            (u'13 Ago, 2014', datetime(2014, 8, 13)),
-            (u'13 Septiembre, 2014', datetime(2014, 9, 13)),
-            (u'13 Setembro, 2014', datetime(2014, 9, 13)),
-        ]
-        parser = AutoDetectLanguage(None, allow_redetection=True)
+    @parameterized.expand([
+        param(date_strings=["11 abril 2010"], expected_language='es'),
+        param(date_strings=["11 junio 2010"], expected_language='es'),
+        param(date_strings=["13 Ago, 2014", "13 Septiembre, 2014"], expected_language='es'),
+    ])
+    def test_do_not_exclude_ineligible_languages_without_modify(self, date_strings, expected_language):
+        self.given_parser(languages=self.known_languages)
+        self.when_one_language_is_detected(date_strings, modify=False)
+        self.then_detected_languages_are([expected_language])
+        self.then_parser_languages_are(self.known_languages)
 
-        for date_string, correct_date in date_fixtures:
-            parsed_date = parser.parse(date_string, None)
-            self.assertEqual(correct_date.date(), parsed_date.date())
+    @parameterized.expand([
+        param(date_strings=["11 abril 2010"], expected_languages=['es', 'pt']),
+        param(date_strings=["11 junio 2010"], expected_languages=['es']),
+        param(date_strings=["13 Ago, 2014", "13 Septiembre, 2014"], expected_languages=['es']),
+        param(date_strings=["13 Srpen, 2014"], expected_languages=['cz']),
+    ])
+    def test_do_not_exclude_ineligible_languages_when_all_ineligible(self, date_strings, expected_languages):
+        self.given_parser(languages=self.known_languages)
+        self.when_all_languages_are_detected(date_strings, modify=True)
+        self.then_detected_languages_are(expected_languages)
+        self.then_parser_languages_are(self.known_languages)
+
+    @parameterized.expand([
+        param(language='es', date_strings=["13 Setembro, 2014"]),
+        param(language='cz', date_strings=["'11 Ağustos, 2014'"]),
+    ])
+    def test_reject_dates_in_other_languages_without_redetection(self, language, date_strings):
+        self.given_parser(languages=self.known_languages)
+        self.given_parser_languages_are([language])
+        self.when_all_languages_are_detected(date_strings)
+        self.then_detected_languages_are([])
+
+    @parameterized.expand([
+        param(detected_languages=['es'], date_strings=['13 Juillet, 2014'], expected_languages=['fr']),
+        param(detected_languages=['es'], date_strings=['11 Ağustos, 2014'], expected_languages=['tr']),
+    ])
+    def test_accept_dates_in_other_languages_with_redetection_enabled(
+        self, detected_languages, date_strings, expected_languages
+    ):
+        self.given_parser(languages=self.known_languages, allow_redetection=True)
+        self.given_parser_languages_are(detected_languages)
+        self.when_all_languages_are_detected(date_strings)
+        self.then_detected_languages_are(expected_languages)
+
+    def test_accept_numeric_dates_without_redetection(self,):
+        self.given_parser(languages=self.known_languages)
+        self.given_parser_languages_are(['es'])
+        self.when_all_languages_are_detected(['13/08/2014'])
+        self.then_detected_languages_are(['es'])
+
+    def given_parser(self, languages=None, allow_redetection=False):
+        if languages is not None:
+            language_map = default_language_loader.get_language_map()
+            languages = [language_map[language]
+                         for language in languages]
+        self.parser = AutoDetectLanguage(languages=languages, allow_redetection=allow_redetection)
+
+    def given_parser_languages_are(self, languages):
+        language_map = default_language_loader.get_language_map()
+        self.parser.languages = [language_map[language]
+                                 for language in languages]
+
+    def when_all_languages_are_detected(self, date_strings, modify=False):
+        assert not isinstance(date_strings, basestring)
+        for date_string in date_strings:
+            detected_languages = list(self.parser.iterate_applicable_languages(date_string, modify=modify))
+        self.detected_languages = detected_languages
+
+    def when_one_language_is_detected(self, date_strings, modify=False):
+        for date_string in date_strings:
+            detected_language = self.parser.iterate_applicable_languages(date_string, modify=modify).next()
+        self.detected_languages = [detected_language]
+
+    def then_detected_languages_are(self, expected_languages):
+        shortnames = map(attrgetter('shortname'), self.detected_languages)
+        self.assertItemsEqual(expected_languages, shortnames)
+
+    def then_parser_languages_are(self, expected_languages):
+        shortnames = map(attrgetter('shortname'), self.parser.languages)
+        self.assertItemsEqual(expected_languages, shortnames)
 
 
-class ExactLanguagesTest(unittest.TestCase):
+class ExactLanguagesTest(BaseTestCase):
+    def setUp(self):
+        super(ExactLanguagesTest, self).setUp()
+        self.parser = NotImplemented
+        self.detected_languages = NotImplemented
 
-    def test_force_setting_language(self):
-        with self.assertRaisesRegexp(TypeError, 'takes exactly 2 arguments'):
-            ExactLanguages()
+    def test_languages_passed_in_constructor_should_not_be_none(self):
+        self.when_parser_is_constructed(languages=None)
+        self.then_error_was_raised(ValueError, 'language cannot be None for ExactLanguages')
 
-        with self.assertRaisesRegexp(ValueError, 'cannot be None'):
-            ExactLanguages(None)
+    @parameterized.expand([
+        param(languages=['es'], date_strings=["13 Ago, 2014"]),
+        param(languages=['es'], date_strings=["13 Septiembre, 2014"]),
+        param(languages=['es'], date_strings=["13/03/2014"]),
+        param(languages=['es'], date_strings=["11/03/2014"]),
+    ])
+    def test_parse_date_in_exact_language(self, languages, date_strings):
+        self.given_parser(languages)
+        self.when_languages_are_detected(date_strings)
+        self.then_detected_languages_are(languages)
 
-    @unittest.skip('This test should only be testing detecting languages, not parsing them. Although tests '
-                   'for parsing this dates should be created separatly to not reduce the coverage')
-    def test_parse_date_in_exact_language(self):
-        date_fixtures = [
-            (u'13 Ago, 2014', datetime(2014, 8, 13)),
-            (u'13 Septiembre, 2014', datetime(2014, 9, 13)),
-            (u'13/03/2014', datetime(2014, 3, 13)),
+    @parameterized.expand([
+        param(languages=['es'], date_strings=["13 Setembro, 2014"]),
+    ])
+    def test_reject_dates_in_other_languages(self, languages, date_strings):
+        self.given_parser(languages=languages)
+        self.when_languages_are_detected(date_strings)
+        self.then_detected_languages_are([])
 
-            # TODO: make the following test pass
-            # in this case, it should have detected spanish as the
-            # language, and so it should use d/m/Y instead of d/m/Y
-            # (u'11/03/2014', datetime(2014, 3, 11)),
-        ]
-        spanish = LanguageDataLoader().get_language('es')
-        parser = ExactLanguages([spanish])
+    def given_parser(self, languages):
+        language_map = default_language_loader.get_language_map()
+        languages = [language_map[language]
+                     for language in languages]
+        self.parser = ExactLanguages(languages)
 
-        for date_string, correct_date in date_fixtures:
-            parsed_date = parser.parse(date_string, None)
-            self.assertEqual(correct_date.date(), parsed_date.date())
+    def when_languages_are_detected(self, date_strings, modify=False):
+        assert not isinstance(date_strings, basestring)
+        for date_string in date_strings:
+            detected_languages = list(self.parser.iterate_applicable_languages(date_string, modify=modify))
+        self.detected_languages = detected_languages
 
-        with self.assertRaisesRegexp(ValueError, 'Invalid date'):
-            portuguese_date = u'13 Setembro, 2014'
-            parser.parse(portuguese_date, None)
+    def when_parser_is_constructed(self, languages):
+        try:
+            ExactLanguages(languages)
+        except Exception as error:
+            self.error = error
+
+    def then_detected_languages_are(self, expected_languages):
+        shortnames = map(attrgetter('shortname'), self.detected_languages)
+        self.assertItemsEqual(expected_languages, shortnames)
 
 
 class TestDateParser(BaseTestCase):
     def setUp(self):
         super(TestDateParser, self).setUp()
-        self.date_string = NotImplemented
         self.parser = NotImplemented
         self.result = NotImplemented
         self.date_parser = NotImplemented
@@ -113,15 +202,32 @@ class TestDateParser(BaseTestCase):
         param('November 19, 2014 at noon', datetime(2014, 11, 19, 12, 0)),
         param('December 13, 2014 at midnight', datetime(2014, 12, 13, 0, 0)),
         param('Nov 25 2014 10:17 pm EST', datetime(2014, 11, 26, 3, 17)),
+        param('April 9, 2013 at 6:11 a.m.', datetime(2013, 4, 9, 6, 11)),
+        param('Aug. 9, 2012 at 2:57 p.m.', datetime(2012, 8, 9, 14, 57)),
+        param('December 10, 2014, 11:02:21 pm', datetime(2014, 12, 10, 23, 2, 21)),
+        param('8:25 a.m. Dec. 12, 2014', datetime(2014, 12, 12, 8, 25)),
+        param('2:21 p.m., December 11, 2014', datetime(2014, 12, 11, 14, 21)),
+        param('Fri, 12 Dec 2014 10:55:50', datetime(2014, 12, 12, 10, 55, 50)),
+        param('20 Mar 2013 10h11', datetime(2013, 3, 20, 10, 11)),
+        param('10:06am Dec 11, 2014', datetime(2014, 12, 11, 10, 6)),
+        param('19 February 2013 year 09:10', datetime(2013, 2, 19, 9, 10)),
         # French dates
         param('11 Mai 2014', datetime(2014, 5, 11)),
         param('dimanche, 11 Mai 2014', datetime(2014, 5, 11)),
-        param('22 janvier 2015 \xe0 14h40', datetime(2015, 1, 22, 14, 40)),
-        param('Dimanche 1er F\xe9vrier \xe0 21:24', datetime(2012, 2, 1, 21, 24)),
+        param('22 janvier 2015 à 14h40', datetime(2015, 1, 22, 14, 40)),
+        param('Dimanche 1er Février à 21:24', datetime(2012, 2, 1, 21, 24)),
+        param('vendredi, décembre 5 2014.', datetime(2014, 12, 5, 0, 0)),
+        param('le 08 Déc 2014 15:11', datetime(2014, 12, 8, 15, 11)),
+        param('Le 11 Décembre 2014 à 09:00', datetime(2014, 12, 11, 9, 0)),
+        param('fév 15, 2013', datetime(2013, 2, 15, 0, 0)),
         # Spanish dates
         param('Martes 21 de Octubre de 2014', datetime(2014, 10, 21)),
         param('Miércoles 20 de Noviembre de 2013', datetime(2013, 11, 20)),
         param('12 de junio del 2012', datetime(2012, 6, 12)),
+        param('13 Ago, 2014', datetime(2014, 8, 13)),
+        param('13 Septiembre, 2014', datetime(2014, 9, 13)),
+        param('11 Marzo, 2014', datetime(2014, 3, 11)),
+        param('julio 5, 2015 en 1:04 pm', datetime(2015, 7, 5, 13, 4)),
         # Dutch dates
         param('11 augustus 2014', datetime(2014, 8, 11)),
         param('14 januari 2014', datetime(2014, 1, 14)),
@@ -131,13 +237,16 @@ class TestDateParser(BaseTestCase):
         param('26 gennaio 2014', datetime(2014, 1, 26)),
         # Portuguese dates
         param('sexta-feira, 10 de junho de 2014 14:52', datetime(2014, 6, 10, 14, 52)),
+        param('13 Setembro, 2014', datetime(2014, 9, 13)),
         # Russian dates
         param('10 мая', datetime(2012, 5, 10)),  # forum.codenet.ru
         param('26 апреля', datetime(2012, 4, 26)),
         param('20 ноября 2013', datetime(2013, 11, 20)),
         param('28 октября 2014 в 07:54', datetime(2014, 10, 28, 7, 54)),
         param('13 января 2015 г. в 13:34', datetime(2015, 1, 13, 13, 34)),
+        param('09 августа 2012', datetime(2012, 8, 9, 0, 0)),
         # Turkish dates
+        param('11 Ağustos, 2014', datetime(2014, 8, 11)),
         param('08.Haziran.2014, 11:07', datetime(2014, 6, 8, 11, 7)),  # forum.andronova.net
         param('17.Şubat.2014, 17:51', datetime(2014, 2, 17, 17, 51)),
         param('14-Aralık-2012, 20:56', datetime(2012, 12, 14, 20, 56)),  # forum.ceviz.net
@@ -152,6 +261,8 @@ class TestDateParser(BaseTestCase):
         param('18.10.14 um 22:56 Uhr', datetime(2014, 10, 18, 22, 56)),
         # Czech dates
         param('pon 16. čer 2014 10:07:43', datetime(2014, 6, 16, 10, 7, 43)),
+        param('13 Srpen, 2014', datetime(2014, 8, 13)),
+        param('čtv 14. lis 2013 12:38:43', datetime(2013, 11, 14, 12, 38, 43)),
         # Thai dates
         param('ธันวาคม 11, 2014, 08:55:08 PM', datetime(2014, 12, 11, 20, 55, 8)),
         param('22 พฤษภาคม 2012, 22:12', datetime(2012, 5, 22, 22, 12)),
@@ -165,30 +276,14 @@ class TestDateParser(BaseTestCase):
         param('05 Tháng một 2015 - 03:54 AM', datetime(2015, 1, 5, 3, 54)),
         # Numeric dates
         param('06-17-2014', datetime(2014, 6, 17)),
-        # Miscellaneous
-        param('April 9, 2013 at 6:11 a.m.', datetime(2013, 4, 9, 6, 11)),
-        param('Aug. 9, 2012 at 2:57 p.m.', datetime(2012, 8, 9, 14, 57)),
-        param('December 10, 2014, 11:02:21 pm', datetime(2014, 12, 10, 23, 2, 21)),
-        param('vendredi, d\xe9cembre 5 2014.', datetime(2014, 12, 5, 0, 0)),
-        param('le 08 D\xe9c 2014 15:11', datetime(2014, 12, 8, 15, 11)),
-        param('Le 11 D\xe9cembre 2014 \xe0 09:00', datetime(2014, 12, 11, 9, 0)),
-        param('f\xe9v 15, 2013', datetime(2013, 2, 15, 0, 0)),
-        param('8:25 a.m. Dec. 12, 2014', datetime(2014, 12, 12, 8, 25)),
-        param('2:21 p.m., December 11, 2014', datetime(2014, 12, 11, 14, 21)),
+        param('13/03/2014', datetime(2014, 3, 13)),
         param('11. 12. 2014, 08:45:39', datetime(2014, 11, 12, 8, 45, 39)),
-        param('Fri, 12 Dec 2014 10:55:50', datetime(2014, 12, 12, 10, 55, 50)),
-        param('čtv 14. lis 2013 12:38:43', datetime(2013, 11, 14, 12, 38, 43)),
-        param('09 августа 2012', datetime(2012, 8, 9, 0, 0)),
-        param('20 Mar 2013 10h11', datetime(2013, 3, 20, 10, 11)),
-        param('10:06am Dec 11, 2014', datetime(2014, 12, 11, 10, 6)),
-        param('19 February 2013 year 09:10', datetime(2013, 2, 19, 9, 10)),
     ])
     def test_dates_parsing(self, date_string, expected):
         self.given_utcnow(datetime(2012, 11, 13))  # Tuesday
         self.given_local_tz_offset(0)
         self.given_parser()
-        self.given_date_string(date_string)
-        self.when_date_is_parsed()
+        self.when_date_is_parsed(date_string)
         self.then_date_was_parsed_by_date_parser()
         self.then_period_is('day')
         self.then_date_obj_exactly_is(expected)
@@ -198,31 +293,33 @@ class TestDateParser(BaseTestCase):
         param('17th October, 2034 @ 01:08 am PDT', datetime(2034, 10, 17, 9, 8)),
         param('15 May 2004 23:24 EDT', datetime(2004, 5, 16, 4, 24)),
         param('15 May 2004', datetime(2004, 5, 15, 0, 0)),
-        ])
+    ])
     def test_parsing_with_time_zones(self, date_string, expected):
         self.given_local_tz_offset(+1)
         self.given_parser()
-        self.given_date_string(date_string)
-        self.when_date_is_parsed()
+        self.when_date_is_parsed(date_string)
         self.then_date_was_parsed_by_date_parser()
         self.then_period_is('day')
         self.then_date_obj_exactly_is(expected)
 
+    def test_empty_dates_string_is_not_parsed(self):
+        self.when_date_is_parsed_by_date_parser('')
+        self.then_error_was_raised(ValueError, "Empty string")
+
     @parameterized.expand([
-        param(''),
         param('invalid date string'),
         param('Aug 7, 2014Aug 7, 2014'),
+        param('24h ago'),
     ])
     def test_dates_not_parsed(self, date_string):
-        self.given_parser()
-        self.given_date_string(date_string)
-        self.when_date_is_parsed()
-        self.then_date_was_not_parsed()
+        self.when_date_is_parsed_by_date_parser(date_string)
+        self.then_error_was_raised(ValueError, "unknown string format")
 
     @parameterized.expand([
         param('10 December', datetime(2014, 12, 10)),
         param('March', datetime(2014, 3, 15)),
         param('Friday', datetime(2015, 2, 13)),
+        param('Monday', datetime(2015, 2, 9)),
         param('10:00PM', datetime(2015, 2, 14, 22, 00)),
         param('16:10', datetime(2015, 2, 14, 16, 10)),
         param('14:05', datetime(2015, 2, 15, 14, 5)),
@@ -232,16 +329,15 @@ class TestDateParser(BaseTestCase):
         self.given_utcnow(datetime(2015, 2, 15, 15, 30))  # Sunday
         self.given_local_tz_offset(0)
         self.given_parser()
-        self.given_date_string(date_string)
-        self.when_date_is_parsed()
+        self.when_date_is_parsed(date_string)
         self.then_date_was_parsed_by_date_parser()
-        self.then_period_is('day')
         self.then_date_obj_exactly_is(expected)
 
     @parameterized.expand([
         param('10 December', datetime(2015, 12, 10)),
         param('March', datetime(2015, 3, 15)),
         param('Friday', datetime(2015, 2, 20)),
+        param('Monday', datetime(2015, 2, 16)),
         param('10:00PM', datetime(2015, 2, 15, 22, 00)),
         param('16:10', datetime(2015, 2, 15, 16, 10)),
         param('14:05', datetime(2015, 2, 16, 14, 5)),
@@ -251,10 +347,8 @@ class TestDateParser(BaseTestCase):
         self.given_utcnow(datetime(2015, 2, 15, 15, 30))  # Sunday
         self.given_local_tz_offset(0)
         self.given_parser()
-        self.given_date_string(date_string)
-        self.when_date_is_parsed()
+        self.when_date_is_parsed(date_string)
         self.then_date_was_parsed_by_date_parser()
-        self.then_period_is('day')
         self.then_date_obj_exactly_is(expected)
 
     @parameterized.expand([
@@ -270,10 +364,8 @@ class TestDateParser(BaseTestCase):
         self.given_utcnow(datetime(2015, 2, 15, 15, 30))  # Sunday
         self.given_local_tz_offset(0)
         self.given_parser()
-        self.given_date_string(date_string)
-        self.when_date_is_parsed()
+        self.when_date_is_parsed(date_string)
         self.then_date_was_parsed_by_date_parser()
-        self.then_period_is('day')
         self.then_date_obj_exactly_is(expected)
 
     @parameterized.expand([
@@ -288,8 +380,7 @@ class TestDateParser(BaseTestCase):
         self.given_configuration('PREFER_DAY_OF_MONTH', 'current')
         self.given_utcnow(today)
         self.given_parser()
-        self.given_date_string(date_string)
-        self.when_date_is_parsed()
+        self.when_date_is_parsed(date_string)
         self.then_date_was_parsed_by_date_parser()
         self.then_date_obj_exactly_is(expected)
 
@@ -305,8 +396,7 @@ class TestDateParser(BaseTestCase):
         self.given_configuration('PREFER_DAY_OF_MONTH', 'last')
         self.given_utcnow(today)
         self.given_parser()
-        self.given_date_string(date_string)
-        self.when_date_is_parsed()
+        self.when_date_is_parsed(date_string)
         self.then_date_was_parsed_by_date_parser()
         self.then_date_obj_exactly_is(expected)
 
@@ -322,8 +412,7 @@ class TestDateParser(BaseTestCase):
         self.given_configuration('PREFER_DAY_OF_MONTH', 'first')
         self.given_utcnow(today)
         self.given_parser()
-        self.given_date_string(date_string)
-        self.when_date_is_parsed()
+        self.when_date_is_parsed(date_string)
         self.then_date_was_parsed_by_date_parser()
         self.then_date_obj_exactly_is(expected)
 
@@ -336,8 +425,7 @@ class TestDateParser(BaseTestCase):
         self.given_configuration('PREFER_DAY_OF_MONTH', prefer_day_of_month)
         self.given_utcnow(datetime(2015, 2, 12))
         self.given_parser()
-        self.given_date_string('24 April 2012')
-        self.when_date_is_parsed()
+        self.when_date_is_parsed('24 April 2012')
         self.then_date_was_parsed_by_date_parser()
         self.then_date_obj_exactly_is(datetime(2012, 4, 24))
 
@@ -349,8 +437,29 @@ class TestDateParser(BaseTestCase):
         param('31 September 2015'),
     ])
     def test_error_should_be_raised_for_invalid_dates_with_too_large_day_number(self, date_string):
-        with self.assertRaisesRegexp(ValueError, 'Day not in range for month'):
-            DateParser().parse(date_string)
+        self.when_date_is_parsed_by_date_parser(date_string)
+        self.then_error_was_raised(ValueError, 'Day not in range for month')
+
+    @parameterized.expand([
+        param('10 December', expected=datetime(2015, 12, 10), period='day'),
+        param('March', expected=datetime(2015, 3, 15), period='month'),
+        param('April', expected=datetime(2015, 4, 15), period='month'),
+        param('December', expected=datetime(2015, 12, 15), period='month'),
+        param('Friday', expected=datetime(2015, 2, 13), period='day'),
+        param('Monday', expected=datetime(2015, 2, 9), period='day'),
+        param('10:00PM', expected=datetime(2015, 2, 15, 22, 00), period='day'),
+        param('16:10', expected=datetime(2015, 2, 15, 16, 10), period='day'),
+        param('2014', expected=datetime(2014, 2, 15), period='year'),
+        param('2008', expected=datetime(2008, 2, 15), period='year'),
+    ])
+    def test_extracted_period(self, date_string, expected=None, period=None):
+        self.given_utcnow(datetime(2015, 2, 15, 15, 30))  # Sunday
+        self.given_local_tz_offset(0)
+        self.given_parser()
+        self.when_date_is_parsed(date_string)
+        self.then_date_was_parsed_by_date_parser()
+        self.then_date_obj_exactly_is(expected)
+        self.then_period_is(period)
 
     def given_utcnow(self, now):
         datetime_mock = Mock(wraps=datetime)
@@ -363,9 +472,6 @@ class TestDateParser(BaseTestCase):
                          'local_tz_offset',
                          new=timedelta(seconds=3600 * offset))
         )
-
-    def given_date_string(self, date_string):
-        self.date_string = date_string
 
     def given_parser(self):
         def collecting_get_date_data(parse):
@@ -385,8 +491,14 @@ class TestDateParser(BaseTestCase):
     def given_configuration(self, key, value):
         self.add_patch(patch.object(settings, key, new=value))
 
-    def when_date_is_parsed(self):
-        self.result = self.parser.get_date_data(self.date_string)
+    def when_date_is_parsed(self, date_string):
+        self.result = self.parser.get_date_data(date_string)
+
+    def when_date_is_parsed_by_date_parser(self, date_string):
+        try:
+            self.result = DateParser().parse(date_string)
+        except Exception as error:
+            self.error = error
 
     def then_period_is(self, period):
         self.assertEqual(period, self.result['period'])
@@ -394,106 +506,8 @@ class TestDateParser(BaseTestCase):
     def then_date_obj_exactly_is(self, expected):
         self.assertEqual(expected, self.result['date_obj'])
 
-    def then_date_was_not_parsed(self):
-        self.assertIsNone(self.result['date_obj'], '"%s" should not be parsed' % self.date_string)
-
     def then_date_was_parsed_by_date_parser(self):
-        self.assertEqual(self.result['date_obj'], self.date_result)
-
-
-@unittest.skip('There are mostly old language detection tests left. New tests should be written.')
-class TestDateParser_(BaseTestCase):
-
-    @unittest.skip('DateParser not using formats anymore. Should be tested separately.')
-    def test_it_dates_with_format(self):
-        parser = DateParser()
-        date = parser.parse('14 giu 13', date_format='%y %B %d')
-        self.assertEqual(date.year, 2014)
-        self.assertEqual(date.month, 6)
-        self.assertEqual(date.day, 13)
-
-        date = parser.parse('14_luglio_15', date_format='%y_%b_%d')
-        self.assertEqual(date.year, 2014)
-        self.assertEqual(date.month, 7)
-        self.assertEqual(date.day, 15)
-
-        date = parser.parse('14_LUGLIO_15', date_format='%y_%b_%d')
-        self.assertEqual(date.year, 2014)
-        self.assertEqual(date.month, 7)
-        self.assertEqual(date.day, 15)
-
-    def test_premature_detection(self):
-        invalid_date_string = '24h ago'  # 'ago' is shortened august in some languages
-        with self.assertRaisesRegexp(ValueError, 'Invalid date: {}'.format(invalid_date_string)):
-            DateParser().parse(invalid_date_string)
-
-    def test_should_not_assume_language_prematurely(self):
-        dp = DateParser()
-        date_fixtures = [
-            (u'07/07/2014', datetime(2014, 7, 7)),
-            (u'07.ago.2014 | 12:52', datetime(2014, 8, 7)),
-            (u'07.jul.2014 | 12:52', datetime(2014, 7, 7)),
-            (u'07.feb.2014 | 12:52', datetime(2014, 2, 7)),
-            (u'07.ene.2014 | 12:52', datetime(2014, 1, 7)),
-        ]
-
-        for date_string, correct_date in date_fixtures:
-            self.assertEqual(correct_date.date(), dp.parse(date_string).date())
-
-    def test_should_not_allow_multiple_languages_by_default(self):
-        dates_in_spanish = [
-            (u'13 Ago, 2014', datetime(2014, 8, 13)),
-            (u'11 Marzo, 2014', datetime(2014, 3, 11)),
-            (u'13 Septiembre, 2014', datetime(2014, 9, 13)),
-        ]
-        dp = DateParser()
-
-        for date_string, correct_date in dates_in_spanish:
-            parsed_date = dp.parse(date_string)
-            self.assertEqual(correct_date.date(), parsed_date.date())
-
-        with self.assertRaisesRegexp(ValueError, 'Invalid date'):
-            portuguese_date = u'13 Setembro, 2014'
-            dp.parse(portuguese_date)
-
-    def test_should_enable_redetection_for_multiple_languages(self):
-        dates_fixture = [
-            (u'13 Ago, 2014', datetime(2014, 8, 13)),
-            (u'11 Marzo, 2014', datetime(2014, 3, 11)),
-            (u'13 Septiembre, 2014', datetime(2014, 9, 13)),
-            (u'13 Setembro, 2014', datetime(2014, 9, 13)),
-            (u'13 Março, 2014', datetime(2014, 3, 13)),
-        ]
-        dp = DateParser(allow_redetect_language=True)
-
-        for date_string, correct_date in dates_fixture:
-            parsed_date = dp.parse(date_string)
-            self.assertEqual(correct_date.date(), parsed_date.date())
-
-    def test_finding_no_language_when_not_seen_before_should_raise_error(self):
-        """ This test depends on the defined order of the languages and needs to be rewritten """
-        dp = DateParser()
-        self.assertEqual(datetime(2014, 8, 13).date(),
-                         dp.parse("13 Srpen, 2014").date())
-
-        with self.assertRaises(ValueError):
-            dp.parse('11 Ağustos, 2014')
-
-    def test_finding_no_language_when_enabled_should_redetect(self):
-        dp = DateParser(allow_redetect_language=True)
-        self.assertEqual(datetime(2014, 8, 13).date(),
-                         dp.parse('13 Ago, 2014').date())
-
-        self.assertEqual(datetime(2014, 8, 11).date(),
-                         dp.parse(u'11 Ağustos, 2014').date())
-
-    def test_finding_no_language_should_work_for_numeric_dates(self):
-        dp = DateParser()
-        self.assertEqual(datetime(2014, 8, 13).date(),
-                         dp.parse('13 Ago, 2014').date())
-
-        parsed_date = dp.parse(u'13/08/2014')
-        self.assertEqual(datetime(2014, 8, 13).date(), parsed_date.date())
+        self.assertEqual(self.result['date_obj'], self.date_result[0])
 
 
 if __name__ == '__main__':