+++ HISTORY.rst
@@ -1,9 +1,24 @@
 .. :changelog:
 
 History
--------
+=======
+
+
+0.2.0 (2015-06-17)
+------------------
+* Easy to use `parse` function
+* Languages definitions using YAML.
+* Using translation based approach for parsing non-english languages. Previously, :mod:`dateutil.parserinfo` was used for language definitions.
+* Better period extraction.
+* Improved tests.
+* Added a number of new simplifications for more comprehensive generic parsing.
+* Improved validation for dates.
+* Support for Polish, Thai and Arabic dates.
+* Support for :mod:`pytz` timezones.
+* Fixed building and packaging issues.
+
 
 0.1.0 (2014-11-24)
----------------------
+------------------
 
 * First release on PyPI.